refactor: move isConstantName to separate file and remove unnecessary test
- Move isConstantName from utils.ts to dedicated isConstantName.ts file
- Remove PascalCase test as variable names can only start with lowercase
- Update imports in index.ts and test file
