+++ packages/syntax-rules/src/isConstantName.ts
